+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 nylas-python Changelog
 ======================
 
+Unreleased (dev)
+----------------
+* You can now pass a list of `scopes` when calling `APIClient.authentication_url()`
+  in order to enable
+  [selective sync](https://docs.nylas.com/docs/how-to-use-selective-sync).
+  Previously, we only set `scope=email` by default; now, the default is to use
+  all scopes.
+
 v4.4.0
 ------
 * Add support for `revoke-all` endpoint.