+++ CHANGELOG.md
@@ -3,6 +3,10 @@ nylas-python Changelog
 
 Unreleased (dev)
 ----------------
+nothing yet
+
+v4.6.0
+------
 * You can now pass a list of `scopes` when calling `APIClient.authentication_url()`
   in order to enable
   [selective sync](https://docs.nylas.com/docs/how-to-use-selective-sync).